Added ostream<< for pairs of printable things and tests for the new operator
One should be able to run "is" with any of the basic standard library types.
However, is needs to be able to print the expected value.  This enables that
printing.  It is also safe for others' code since if they need to use their own
operator<<(ostream,pair), ours is hidden safely away in the TAP namespace.